devices/truck_gps_tracker/requirements.txt
pyserial==3.5
pynmea2==1.19.0
httpx==0.27.2

